Use react native navigation to control modal visibility

+++ app/containers/editHabbajetScreen.tsx
@@ -1,4 +1,4 @@
-import React, {useState} from 'react';
+import React from 'react';
 import {useDispatch, useSelector} from 'react-redux';
 import {
     addHabbajetAction,
@@ -17,8 +17,7 @@ import {
     getSelectedHabbajet,
 } from '../selectors';
 import HabbajetForm from '../components/habbajetForm';
-import ConfirmationModal from '../components/confirmationModal';
-import {View} from 'react-native';
+import {Navigation} from 'react-native-navigation';
 
 const EditHabbajetScreen = () => {
     const dispatch = useDispatch();
@@ -27,7 +26,6 @@ const EditHabbajetScreen = () => {
     const newHabbajet = useSelector(getValuesForNewHabbajet);
     const selectedHabbajet = useSelector(getSelectedHabbajet);
     const habbajetNames = useSelector(getUnselectedHabbajetNames);
-    const [modalVisibility, setModalVisibility] = useState(false);
 
     if (isValid) {
         dispatch(clearEditorAction());
@@ -46,34 +44,37 @@ const EditHabbajetScreen = () => {
     }
 
     return (
-        <View>
-            <ConfirmationModal
-                text={'Are you sure you want to delete this habbajet?'}
-                visible={modalVisibility}
-                onConfirm={() => {
-                    goToHome();
-                    dispatch(deleteHabbajetAction());
-                    saveState();
-                }}
-                onDismiss={() => {
-                    setModalVisibility(false);
-                }}
-            />
-            <HabbajetForm
-                nameField={fields.name}
-                valueField={fields.value}
-                modifierField={fields.modifier}
-                slackField={fields.slack}
-                selectedColor={newHabbajet.color}
-                onUpdate={(key, value) =>
-                    dispatch(updateEditorFieldAction(key, value))
-                }
-                onSubmit={() => {
-                    dispatch(validateEditorAction(habbajetNames));
-                }}
-                onDelete={() => setModalVisibility(true)}
-            />
-        </View>
+        <HabbajetForm
+            nameField={fields.name}
+            valueField={fields.value}
+            modifierField={fields.modifier}
+            slackField={fields.slack}
+            selectedColor={newHabbajet.color}
+            onUpdate={(key, value) =>
+                dispatch(updateEditorFieldAction(key, value))
+            }
+            onSubmit={() => {
+                dispatch(validateEditorAction(habbajetNames));
+            }}
+            onDelete={() =>
+                Navigation.showModal({
+                    component: {
+                        name: 'modal.confirmation',
+                        id: 'deleteModal',
+                        passProps: {
+                            id: 'deleteModal',
+                            text:
+                                'Are you sure you want to delete this habbajet?',
+                            onConfirm: () => {
+                                goToHome();
+                                dispatch(deleteHabbajetAction());
+                                saveState();
+                            },
+                        },
+                    },
+                })
+            }
+        />
     );
 };
